+++ red_lion/main.tf
@@ -1,10 +1,8 @@
 #############################################################################
-## Consul Open Source
+# Consul Open Source
 #############################################################################
 
-###
-### Consul module outputs
-###
+# Consul module outputs
 
 output "consul_oss_server_ips" {
   description = "Consul OSS Server IP addresses"
@@ -16,29 +14,12 @@ output "consul_oss_server_ips" {
   ]
 }
 
-#
-#output "consul_oss_client_ips" {
-#  description = "Consul OSS Client IP addresses"
-#  value = [
-#    "${docker_container.consul_oss_client_0.ip_address}",
-#    "${docker_container.consul_oss_client_1.ip_address}",
-#    "${docker_container.consul_oss_client_2.ip_address}"
-#  ]
-#}
-
-#output "consul_oss_server_ips" {
-#  description = "Consul OSS Server IP addresses"
-#  value = ["${docker_container.consul_oss_server.*.ip_address}"]
-#}
-
 output "consul_oss_client_ips" {
   description = "Consul OSS Client IP addresses"
   value       = ["${docker_container.consul_oss_client.*.ip_address}"]
 }
 
-###
-### Consul related variables
-###
+# Consul related variables
 
 variable "consul_log_level" {}
 variable "datacenter_name" {}
@@ -54,19 +35,15 @@ variable "consul_custom_instance_count" {}
 variable "consul_oss" {}
 variable "consul_oss_instance_count" {}
 
-###
-### This is the official Consul Docker image that Vaultron uses by default.
-### See also: https://hub.docker.com/_/consul/
-###
+# This is the official Consul Docker image that Vaultron uses by default.
+# See also: https://hub.docker.com/_/consul/
 
 resource "docker_image" "consul" {
   name         = "consul:${var.consul_version}"
   keep_locally = true
 }
 
-###
-### Consul Open Source server common configuration
-###
+# Consul Open Source server common configuration
 
 data "template_file" "consul_oss_server_common_config" {
   count    = "${var.consul_oss}"
@@ -85,9 +62,39 @@ data "template_file" "consul_oss_server_common_config" {
   }
 }
 
-###
-### Consul Open Source Server 1
-###
+# Consul Server TLS CA Bundle
+
+#data "template_file" "consul_ca_bundle" {
+#  template = "${file("${path.module}/tls/ca-bundle.crt")}"
+#}
+
+# Consul Server TLS certificates and keys
+
+#data "template_file" "consul_server_0_tls_cert" {
+#  template = "${file("${path.module}/tls/consul-server-0.crt")}"
+#}
+#
+#data "template_file" "consul_server_1_tls_cert" {
+#  template = "${file("${path.module}/tls/consul-server-1.crt")}"
+#}
+#
+#data "template_file" "consul_server_2_tls_cert" {
+#  template = "${file("${path.module}/tls/consul-server-2.crt")}"
+#}
+#
+#data "template_file" "consul_server_0_tls_key" {
+#  template = "${file("${path.module}/tls/consul-server-0.key")}"
+#}
+#
+#data "template_file" "consul_server_1_tls_key" {
+#  template = "${file("${path.module}/tls/consul-server-1.key")}"
+#}
+#
+#data "template_file" "consul_server_2_tls_key" {
+#  template = "${file("${path.module}/tls/consul-server-2.key")}"
+#}
+#
+# Consul Open Source Server 1
 
 resource "docker_container" "consul_oss_server_0" {
   count = "${var.consul_oss}"
@@ -100,10 +107,26 @@ resource "docker_container" "consul_oss_server_0" {
   # log_opts = {
   #   gelf-address = "udp://${var.log_server_ip}:5114"
   # }
+
   upload = {
     content = "${data.template_file.consul_oss_server_common_config.rendered}"
     file    = "/consul/config/common_config.json"
   }
+
+#  upload = {
+#    content = "${data.template_file.consul_ca_bundle.rendered}"
+#    file    = "/consul/config/ca-bundle.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_server_0_tls_cert.rendered}"
+#    file    = "/consul/config/consul-server.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_server_0_tls_key.rendered}"
+#    file    = "/consul/config/consul-server.key"
+#  }
 
   volumes {
     host_path      = "${path.module}/../../../consul/consul_oss_server_0/config"
@@ -177,9 +200,7 @@ resource "docker_container" "consul_oss_server_0" {
   }
 }
 
-###
-### Consul Open Source Server 2
-###
+# Consul Open Source Server 2
 
 resource "docker_container" "consul_oss_server_1" {
   count = "${var.consul_oss}"
@@ -195,6 +216,21 @@ resource "docker_container" "consul_oss_server_1" {
     content = "${data.template_file.consul_oss_server_common_config.rendered}"
     file    = "/consul/config/common_config.json"
   }
+
+#  upload = {
+#    content = "${data.template_file.consul_ca_bundle.rendered}"
+#    file    = "/consul/config/ca-bundle.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_server_1_tls_cert.rendered}"
+#    file    = "/consul/config/consul-server.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_server_1_tls_key.rendered}"
+#    file    = "/consul/config/consul-server.key"
+#  }
 
   volumes {
     host_path      = "${path.module}/../../../consul/consul_oss_server_1/config"
@@ -218,9 +254,7 @@ resource "docker_container" "consul_oss_server_1" {
   must_run = true
 }
 
-###
-### Consul Open Source Server 3
-###
+# Consul Open Source Server 3
 
 resource "docker_container" "consul_oss_server_2" {
   count = "${var.consul_oss}"
@@ -236,6 +270,21 @@ resource "docker_container" "consul_oss_server_2" {
     content = "${data.template_file.consul_oss_server_common_config.rendered}"
     file    = "/consul/config/common_config.json"
   }
+
+#  upload = {
+#    content = "${data.template_file.consul_ca_bundle.rendered}"
+#    file    = "/consul/config/ca-bundle.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_server_2_tls_cert.rendered}"
+#    file    = "/consul/config/consul-server.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_server_2_tls_key.rendered}"
+#    file    = "/consul/config/consul-server.key"
+#  }
 
   volumes {
     host_path      = "${path.module}/../../../consul/consul_oss_server_2/config"
@@ -259,9 +308,7 @@ resource "docker_container" "consul_oss_server_2" {
   must_run = true
 }
 
-###
-### Consul Open Source client common configuration
-###
+# Consul Open Source client common configuration
 
 data "template_file" "consul_oss_client_common_config" {
   count    = "${var.consul_oss}"
@@ -272,9 +319,17 @@ data "template_file" "consul_oss_client_common_config" {
   }
 }
 
-###
-### Consul Open Source Clients
-###
+# Consul Open Source Clients
+
+# Consul Client TLS certificates and keys
+
+#data "template_file" "consul_client_tls_cert" {
+#  template = "${file("${path.module}/tls/consul-client-0.crt")}"
+#}
+#
+#data "template_file" "consul_client_tls_key" {
+#  template = "${file("${path.module}/tls/consul-client-0.key")}"
+#}
 
 resource "docker_container" "consul_oss_client" {
   count = "${var.consul_oss_instance_count}"
@@ -285,6 +340,21 @@ resource "docker_container" "consul_oss_client" {
     content = "${data.template_file.consul_oss_client_common_config.rendered}"
     file    = "/consul/config/common_config.json"
   }
+
+#  upload = {
+#    content = "${data.template_file.consul_ca_bundle.rendered}"
+#    file    = "/consul/config/ca-bundle.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_client_tls_cert.rendered}"
+#    file    = "/consul/config/consul-client.crt"
+#  }
+#
+#  upload = {
+#    content = "${data.template_file.consul_client_tls_key.rendered}"
+#    file    = "/consul/config/consul-client.key"
+#  }
 
   volumes {
     host_path      = "${path.module}/../../../consul/consul_oss_client_${count.index}/config"
@@ -309,273 +379,6 @@ resource "docker_container" "consul_oss_client" {
                      )}"]
 
   dns        = ["${docker_container.consul_oss_server_0.ip_address}", "${docker_container.consul_oss_server_1.ip_address}", "${docker_container.consul_oss_server_2.ip_address}"]
-  dns_search = ["consul"]
-  must_run   = true
-}
-
-#############################################################################
-## Consul Custom build
-#############################################################################
-
-###
-### Consul Custom server common configuration
-###
-
-data "template_file" "consul_custom_server_common_config" {
-  count    = "${var.consul_custom}"
-  template = "${file("${path.module}/templates/consul_custom_server_config_${var.consul_version}.tpl")}"
-
-  vars {
-    acl_datacenter   = "arus"
-    bootstrap_expect = 3
-    datacenter       = "${var.datacenter_name}"
-    data_dir         = "${var.consul_data_dir}"
-    client           = "0.0.0.0"
-    recursor1        = "${var.consul_recursor_1}"
-    recursor2        = "${var.consul_recursor_2}"
-    ui               = "true"
-  }
-}
-
-###
-### Consul Custom Server 1
-###
-
-resource "docker_container" "consul_custom_server_0" {
-  count = "${var.consul_custom}"
-  name  = "consul_custom_server_0"
-  env   = ["CONSUL_ALLOW_PRIVILEGED_PORTS="]
-  image = "${docker_image.consul.latest}"
-
-  # TODO: make GELF logging a conditional thing
-  # log_driver = "gelf"
-  # log_opts = {
-  #   gelf-address = "udp://${var.log_server_ip}:5114"
-  # }
-  upload = {
-    content = "${data.template_file.consul_custom_server_common_config.rendered}"
-    file    = "/consul/config/common_config.json"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../custom/"
-    container_path = "/consul/custom"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_server_0/config"
-    container_path = "/consul/config"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_server_0/data"
-    container_path = "/consul/data"
-  }
-
-  entrypoint = ["/consul/custom/consul",
-    "agent",
-    "-server",
-    "-config-dir=/consul/config",
-    "-node=consul_custom_server_0",
-    "-client=0.0.0.0",
-    "-dns-port=53",
-  ]
-
-  must_run = true
-
-  # Define some published ports here for the purpose of connecting into
-  # the cluster from the host system:
-  ports {
-    internal = "8300"
-    external = "8300"
-    protocol = "tcp"
-  }
-
-  ports {
-    internal = "8301"
-    external = "8301"
-    protocol = "tcp"
-  }
-
-  ports {
-    internal = "8301"
-    external = "8301"
-    protocol = "udp"
-  }
-
-  ports {
-    internal = "8302"
-    external = "8302"
-    protocol = "tcp"
-  }
-
-  ports {
-    internal = "8302"
-    external = "8302"
-    protocol = "udp"
-  }
-
-  ports {
-    internal = "8500"
-    external = "8500"
-    protocol = "tcp"
-  }
-
-  ports {
-    internal = "53"
-    external = "8600"
-    protocol = "tcp"
-  }
-
-  ports {
-    internal = "53"
-    external = "8600"
-    protocol = "udp"
-  }
-}
-
-###
-### Consul Custom Server 2
-###
-
-resource "docker_container" "consul_custom_server_1" {
-  count = "${var.consul_custom}"
-  name  = "consul_custom_server_1"
-  image = "${docker_image.consul.latest}"
-
-  # TODO: make GELF logging a conditional thing
-  # log_driver = "gelf"
-  # log_opts = {
-  #   gelf-address = "udp://${var.log_server_ip}:5114"
-  # }
-  upload = {
-    content = "${data.template_file.consul_custom_server_common_config.rendered}"
-    file    = "/consul/config/common_config.json"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../custom/"
-    container_path = "/consul/custom"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_server_1/config"
-    container_path = "/consul/config"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_server_1/data"
-    container_path = "/consul/data"
-  }
-
-  entrypoint = ["/consul/custom/consul",
-    "agent",
-    "-server",
-    "-config-dir=/consul/config",
-    "-node=consul_custom_server_1",
-    "-retry-join=${docker_container.consul_custom_server_0.ip_address}",
-    "-dns-port=53",
-  ]
-
-  must_run = true
-}
-
-###
-### Consul Custom Server 3
-###
-
-resource "docker_container" "consul_custom_server_2" {
-  count = "${var.consul_custom}"
-  name  = "consul_custom_server_2"
-  image = "${docker_image.consul.latest}"
-
-  # TODO: make GELF logging a conditional thing
-  # log_driver = "gelf"
-  # log_opts = {
-  #   gelf-address = "udp://${var.log_server_ip}:5114"
-  # }
-  upload = {
-    content = "${data.template_file.consul_custom_server_common_config.rendered}"
-    file    = "/consul/config/common_config.json"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../custom/"
-    container_path = "/consul/custom"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_server_2/config"
-    container_path = "/consul/config"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_server_2/data"
-    container_path = "/consul/data"
-  }
-
-  entrypoint = ["/consul/custom/consul",
-    "agent",
-    "-server",
-    "-config-dir=/consul/config",
-    "-node=consul_custom_server_2",
-    "-retry-join=${docker_container.consul_custom_server_0.ip_address}",
-    "-dns-port=53",
-  ]
-
-  must_run = true
-}
-
-###
-### Consul Custom client common configuration
-###
-
-data "template_file" "consul_custom_client_common_config" {
-  count    = "${var.consul_custom}"
-  template = "${file("${path.module}/templates/consul_custom_client_config_${var.consul_version}.tpl")}"
-
-  vars {
-    common_configuration = "true"
-  }
-}
-
-###
-### Consul Custom Clients
-###
-
-resource "docker_container" "consul_custom_client" {
-  count = "${var.consul_custom_instance_count}"
-  name  = "${format("consul_custom_client_%d", count.index)}"
-  image = "${docker_image.consul.latest}"
-
-  upload = {
-    content = "${data.template_file.consul_custom_client_common_config.rendered}"
-    file    = "/consul/config/common_config.json"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_client_${count.index}/config"
-    container_path = "/consul/config"
-  }
-
-  volumes {
-    host_path      = "${path.module}/../../../consul/consul_custom_client_${count.index}/data"
-    container_path = "/consul/data"
-  }
-
-  entrypoint = ["${list("consul",
-                     "agent",
-                     "-config-dir=/consul/config",
-                     "-client=0.0.0.0",
-                     "-data-dir=/consul/data",
-                     "-node=consul_custom_client_${count.index}",
-                     "-datacenter=${var.datacenter_name}",
-                     "-retry-join=${docker_container.consul_custom_server_2.ip_address}",
-                     "-retry-join=${docker_container.consul_custom_server_1.ip_address}",
-                     "-retry-join=${docker_container.consul_custom_server_0.ip_address}"
-                     )}"]
-
-  dns        = ["${docker_container.consul_custom_server_0.ip_address}", "${docker_container.consul_custom_server_1.ip_address}", "${docker_container.consul_custom_server_2.ip_address}"]
   dns_search = ["consul"]
   must_run   = true
 }